chore(cmd/influx/write): refactor to csv2lp package
