+++ pkg/csv2lp/csvTable.go
@@ -1,4 +1,4 @@
-package write
+package csv2lp
 
 import (
 	"errors"
