+++ pkg/csv2lp/csvTable_test.go
@@ -1,4 +1,4 @@
-package write
+package csv2lp
 
 import (
 	"encoding/csv"
