+++ pkg/csv2lp/dataConversion.go
@@ -1,4 +1,4 @@
-package write
+package csv2lp
 
 import (
 	"encoding/base64"
